split to per chip files
