+++ firmware/hw_layer/ports/stm32/stm32f7/device_mpu_util.h
@@ -22,3 +22,10 @@
 #define SPI_CR2_24BIT_MODE SPI_CR2_DS_2 | SPI_CR2_DS_1 | SPI_CR2_DS_0
 
 #define ADC_MAX_VALUE 4095
+
+#ifdef __cplusplus
+uintptr_t getBootAddress();
+bool setBootAddress(uintptr_t address);
+void preBootloaderUpdate();
+void postBootloaderUpdate();
+#endif
